AS-REQ without encrypted data only if no preauth set
By default AS-REQ was sent without encrypted data and resent a second time with one of the encryption given in the AS-REP if failing and credential were provided. However, when credentials provided were a certificate and the unencrypted AS-REQ was sent for a principal without PREAUTH, minikerberos failed by trying to decrypt the session key (because it's encrypted with principal's password and not certificate's public key). To fix this and be more RFC compliant, AS-REQ is now sent with encryption if credentials are provided. If the encryption used for authentication data is not supported, it will retry with an encryption method suggested by AS-REP and supported by minikerberos. This also save one request in most of case as we don't have to resend for principals with preauth enabled (which is the case most of the time).